Initialize and open a GLUT window from VisualSim class.
The step() function in the particular case blocks, in order to enter
the glutMainLoop() function, which usually never returns.  (In
freeglut it may return.)  A timed callback is initiated before
entering glutMainLoop() which performs drawing and checks the liblo
server in a non-blocking fashion.

In the future, perhaps drawing could be performed only when messages
are received.