InterfaceSphereFactory forwards creation messages to the other simulations.
The Simulation class gets a list of other simulations it can send
messages to, which is initialized right at the beginning of the
program.  The InterfaceSim subclass, when it receives messages, uses
Simulation::send() to forward the message to all listed simulations.
With this method, all objects will have to be subclassed to know how
to forward each message individually.  Not entirely sure it is the
right approach, but a possibility.