Add a simple enum-based type system to the Simulation class.
While it is not desirable to have information about inherited classes
here, the alternative is to depend on dynamic casting, which is known
to be somewhat slow.  Also, without a value to identify a simulation
type, the sendtotype() function would be more difficult to express.
This function is needed for InterfaceSim to direct messages to
specific simulations.