Specify in the InterfaceSim /get message-forward declarations the simulation type to send to.
This effectively filters incoming /get messages so that they are
responded to by the appropriate simulation type.  It might be
aesthetically cleaner to have each simulation know which messages
respond to (by disabling /get for certain messages for each subclass
for example), but filtering the /get messages before forwarding them
in the first place is a little more efficient, and also requires less
code.