Add support for hinge constraints, as well as all the infrastructure needed for constraints.
This includes the following new classes: HingeFactory,
InterfaceHingeFactory, PhysicsHingeFactory, ODEConstraint,
OscHingeODE, OscHingeInterface.  In addition, OscConstraint gets a
pointer to its parent object just like OscObject (pointing to the
Simulation).  It is set to NULL for older classes.  (If they are
instantiated, it will generate an error.)

Considering the required number of classes for a single joint type,
and that constraints only make sense within the physics context, it
must be decided whether this is the best approach.  However, it seems
good to stay analogous with the class hierarchy for objects, since
constraints, like objects, can respond to external messages.  The nice
thing is that while InterfaceSim gets a copy of the constraint, any
ODE-specific code will be only in the physics simulation's subclass.