Bootstraps, compiles and runs under Windows using MingW/MSYS.
Notes: Questionable path trickery and popen-related command piping was
required to get Python (for SCons) to find the locations of the
libraries and sources.  The SConstruct file contains some finicky
OS-dependent library dependency stuff, which should probably be made
cleaner.  For pthreads, the pthread_t data structure in pthreads-win32
is not a simple integer-sized type, but a data structure.  Therefore,
to check if a thread is running, it was necessary to introduce
Simulation::m_bStarted to use as a boolean instead of checking whether
the pthread_t is NULL.