Fixed bug where c_path() returned a pointer to an internal string which was maintained by a temporary std::string instance.
This caused invalid strings to be passed to Simulation::send().
Solution (and small optimization) is to cache path() and c_path() in
member variable, OscBase::m_path.  This required dropping the const
modifier on c_path(), in turn requiring dropping it for OscValue
callbacks.  Realized it wasn't appropriate there anyways, since
callbacks should be able to modify this referenced object.