Fix virtual function initialiation race condition.
Starting the pthread during the Simulation contructor caused a race
condition for the initialization of virtual functions.  This caused
the subclassed virtual functions to not be called properly, such as
step() and initialize(), leading to segfaults or "pure virtual
function called" errors.

This patch fixes it by moving the pthread code to start() and stop()
functions which are called by the main program after the objects have
been properly initialized.