Introduce a 'density' value for OscObjects.
This enables resetting the mass of spheres and prisms with a constant
density after they are resized and thus fixes the problem of many of
the tests requiring very small forces to succeed.

The 'mass' value still sets the total mass of the object, changing its
density.

This patch also adds a missing /get handler for mass in InterfaceSim,
and makes density also available through InterfaceSim.