Track the grabbed object and apply opposing damped spring forces in PhysicsSim and HapticsSim.
The set_grabbed() function is made a virtual function in simulation()
since it is common functionality here and we can avoid unnecessary
casting.

Still needs work on determining optimal stiffness and damping
coefficients both for the physics engine and for the haptic
controller.  It's complicated because these are a function of the
device inertia and the physics object mass.