Instead of CHAIObject being a base class, it is now a "specialization" member of OscObject.
This works around problems with multiple inheritence, using the
OscValue's set-callback to provide overriding behaviour.  It allows
CHAIObject to take over more of the CHAI-specific operations that are
general to all object types.  One setback is that material-related
functionality cannot be taken care of this way because m_material is
not part of CHAI's cGenericObject class---hopefully this can be
resolved in a future version of CHAI.

In any case, previously the CHAIObject class was next to useless,
merely a container for shared data, due to the impossibility of using
C++ multiple inheritence to override certain behaviour specific to all
CHAI objects.  This lead to duplicated code.  Now instead of
inheriting from CHAIObject, we are "assigning a specialization" to an
OscObject, and the CHAIObject constructor overrides certain OscValue
handlers by setting their callback.