Fix segfault bug in destruction order.
Two problems: the Simulation instances were freeing m_server before
the member OscValues were destroyed, which try to remove methods from
m_server, causing a segfault.  Solution is to manually set m_server =
0 for each OscValue member of a Simulation -- not elegant, but it
works.

Additionally, due to the order in which destructors are called,
calling stop() in ~Simulation() was not sufficient to ensure the
thread has stopped before ~InterfaceSim() et al are called.  Therefore
the thread is running and dereferencing "me" while the object is
partially destroyed.  The solution is to call stop() as early as
possible in the destruction sequence, i.e., in ~InterfaceSim() et al.