Allow to specify remote receivers in addition to local SimulationInfo destinations.
Renames SimulationInfo class to SimulationReceiver, and allows to
instantiate it with a liblo-style URL instead of a Simulation's
lo_address/queue.

Adds possibility of specifying multiple `-s' options on the command
line and following the simulation spec string with a liblo URL.

This means that if Dimple is executed on one computer with "-s v", and
on the other computer with "-s ph -s v,osc.udp://<hostname>:7773",
then the haptics and physics simulations will run on one computer, and
send visual update messages to the remote computer that displays the
graphics.